added api route for fetching github pinned repositories for projects component
