+++ components/Contact.jsx
@@ -2,7 +2,7 @@ import React from 'react'
 
 const Contact = () => {
   return (
-    <div className='h-[65vh] text-7xl font-bold'>
+    <div className='h-[80vh] text-7xl font-bold'>
         Contact
     </div>
   )
